api/requirements.txt
fastapi == 0.108.0
pydantic-settings == 2.1.0
uvicorn == 0.25.0
urllib3==1.26.6 # needed for OpenSSL 1.1.1+
pymongo==4.6.2
websockets==12.0